casting element to float to avoid integer division

+++ src/main/java/ca/corefacility/bioinformatics/irida/model/sample/CoverageQCEntry.java
@@ -82,7 +82,7 @@ public class CoverageQCEntry extends QCEntry {
 	}
 
 	private int calculateCoverage() {
-		return Math.round(totalBases / project.getGenomeSize());
+		return Math.round((float) totalBases / project.getGenomeSize());
 	}
 
 }
